create manitto

+++ manittoWar/src/main/java/edu/fisa/lab/domain/dto/StudentDTO.java
@@ -2,6 +2,7 @@ package edu.fisa.lab.domain.dto;
 
 import edu.fisa.lab.domain.entity.Board;
 import edu.fisa.lab.domain.entity.Student;
+import jakarta.persistence.Column;
 import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -19,7 +20,7 @@ import lombok.ToString;
 @Builder
 public class StudentDTO {
 	 
-	private long id;
+	private Integer id;
 	private String name;
 	private String pw;
 	private long targetId;
